% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot_F_RACE}
\alias{boxplot_F_RACE}
\title{Boxplot of the high housing cost burden percentage vs the most represented race for each county}
\usage{
boxplot_F_RACE(data)
}
\arguments{
\item{data}{SVI_2020_US_county dataset}
}
\value{
A ggplot object with the boxplot
}
\description{
The function returns boxplots of the dataset SVI_2020_US_county according to the formula EP_HBURD ~ F_RACE
}
