% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram_F_RACE}
\alias{histogram_F_RACE}
\title{Histogram of the high housing cost burden percentage vs the most represented race for each county}
\usage{
histogram_F_RACE(data)
}
\arguments{
\item{data}{SVI_2020_US_county data set filtered on the F_RACE label}
}
\value{
A ggplot object with the histogram
}
\description{
The function returns histograms of the data set SVI_2020_US_county according to the formula EP_HBURD ~ F_RACE
}
