% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutational_ANOVA.R
\name{permutational_ANOVA}
\alias{permutational_ANOVA}
\title{Permutational ANOVA of the high housing cost burden percentage vs the most represented race for each county}
\usage{
permutational_ANOVA(data, B = 1000)
}
\arguments{
\item{data}{SVI_2020_US_county dataset filtered on the F_RACE label}

\item{B}{number of permutations}
}
\value{
a list with the T0 statistic, the permutational sample and the pvalue of the test
}
\description{
The function performs a permutational ANOVA of the data set SVI_2020_US_county according to the formula EP_HBURD ~ F_RACE.
The test is
H0: tau1 = tau2 = tau3 = 0 i.e. no group specific effect
H1: (H0)^c i.e there are at least two different populations
}
