% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutational_regression.R
\name{permutational_regression}
\alias{permutational_regression}
\title{Permutational F-type test for regression of the high housing cost burden percentage vs the most represented race for each county}
\usage{
permutational_regression(data, B = 1000)
}
\arguments{
\item{data}{SVI_2020_US_county dataset filtered on the F_RACE label}

\item{B}{number of permutations}
}
\value{
a list with the T0 statistic, the permutational sample and the pvalue of the test
}
\description{
The function performs a permutational F-type test for the linear fit according to the formula EP_HBURD ~ EP_WHITE, EP_AFAM, EP_HISP, EP_ASIAN, EP_NHPI of the data set SVI_2020_US_county.
The test is
H0: beta1 = beta2 = ... = beta5 = 0 i.e. no group specific effect
H1: (H0)^c i.e there is at least one nonzero parameter
}
