% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SVI_2020_US_county}
\alias{SVI_2020_US_county}
\title{Accelerometry Data Resampled from UK Biobank}
\format{
## `SVI_2020_US_county`
A data frame with 3143 rows and 161 columns:
\describe{
}
}
\usage{
SVI_2020_US_county
}
\description{
Toy accelerometry data for BIS620
}
\keyword{datasets}
